Rename server_setup file to server_commands
